pulse/intercept: don't intercept corked streams when limit present
- If a limit is present, any corked streams will be sent directly to the
  queue; only uncorked streams will be dequeued.
- Move the Interceptor, LimitingInterceptor, and InterceptError types
  into the types module, so visibility rules are more consistent.
- Fix a logic issue that prevented bardcast from uncorking itself if one
  of its streams uncorked.
