+++ src/snd/pulse/intercept/mod.rs
@@ -10,10 +10,7 @@ mod types;
 mod util;
 
 use std::borrow::Cow;
-use std::error::Error;
-use std::fmt::{Display, Error as FormatError, Formatter};
 
-use async_trait::async_trait;
 use libpulse_binding::error::Code;
 use log::{debug, error, info, warn};
 use tokio::sync::broadcast::error::RecvError;
@@ -23,6 +20,8 @@ use crate::util::task::ValueJoinHandle;
 use self::types::{
     CapturingInterceptor,
     DuplexingInterceptor,
+    InterceptError,
+    Interceptor,
     SingleInputMonitor,
     QueuedInterceptor
 };
@@ -42,133 +41,6 @@ const TEARDOWN_FAILURE_WARNING: &'static str =
      configuration.";
 const SINK_INPUT_MOVE_FAILURE: &'static str =
     "Some captured inputs failed to move to their original or default sink.";
-
-// TYPE DEFINITIONS ************************************************************
-/// Represents possible error modes for [`Interceptor::record()`].
-#[derive(Debug)]
-enum InterceptError<'a> {
-    /// PulseAudio returned an error while attempting to intercept the
-    /// application.
-    PulseError(Code),
-
-    /// The interceptor cannot concurrently intercept any more applications.
-    AtCapacity(usize, Cow<'a, OwnedSinkInputInfo>),
-}
-
-/// Core abstraction for intercepting application audio.
-///
-/// The main function in this interface is [`intercept`], which performs the
-/// work of intercepting applications and routing them such that bardcast can
-/// read their stream. This interface also provides a few functions to query the
-/// current state of intercepted streams.
-#[async_trait]
-trait Interceptor: Send + Sync {
-    /// Starts recording audio from the given application.
-    async fn record<'a>(
-        &mut self,
-        source: Cow<'a, OwnedSinkInputInfo>
-    ) -> Result<(), InterceptError<'a>>;
-
-    /// Updates the stream metadata for the given intercepted stream. This may
-    /// cork or uncork the stream depending on the cork state of all intercepted
-    /// applications.
-    async fn update_intercept(
-        &mut self,
-        source: &OwnedSinkInputInfo
-    ) -> Result<(), Code>;
-
-    /// Stops recording audio from the given application. If this interceptor
-    /// is not currently recording audio from the given application, this
-    /// returns `Ok(false)`.
-    async fn stop(
-        &mut self,
-        source_idx: u32
-    ) -> Result<bool, Code>;
-
-    /// Determines if [`Interceptor::monitor()`] must be called to ensure
-    /// underlying tasks are properly monitored and cleaned up during normal
-    /// operation.
-    fn needs_monitor(&self) -> bool;
-
-    /// Monitors and cleans up underlying tasks, if any.
-    async fn monitor(&mut self);
-
-    /// Returns the number of applications that have been intercepted by this
-    /// interceptor.
-    fn len(&self) -> usize;
-
-    /// Determines whether the consuming end of the stream has closed.
-    fn stream_closed(&self) -> bool;
-
-    /// Gracefully closes any system resources that were created by the
-    /// `Interceptor`, returning any intercepted streams to another audio device
-    /// if needed.
-    async fn close(mut self);
-}
-
-/// Specialization of [`Interceptor`] that allows for limits on the number of
-/// concurrently captured applications.
-trait LimitingInterceptor: Interceptor {
-    /// Returns the number of concurrent applications that can be intercepted
-    /// by this `LimitedInterceptor` instance.
-    ///
-    /// Returns `None` if there is no limit applied to this interceptor.
-    fn capacity(&self) -> Option<usize>;
-
-    /// Returns the number of additional applications that can be concurrently
-    /// intercepted by this `LimitedInterceptor` instance.
-    ///
-    /// Returns `None` if there is no limit applied to this interceptor.
-    fn remaining(&self) -> Option<usize> {
-        self.capacity().map(|capacity| capacity - self.len())
-    }
-
-    /// Determines whether this interceptor has intercepted the maximum number
-    /// of concurrent applications.
-    ///
-    /// This will always be `false` for interceptors with no limit.
-    fn is_full(&self) -> bool {
-        self.remaining().is_some_and(|remaining| remaining == 0)
-    }
-}
-
-// TRAIT IMPLS *****************************************************************
-impl From<Code> for InterceptError<'_> {
-    fn from(code: Code) -> Self {
-        Self::PulseError(code)
-    }
-}
-
-impl<'a> TryFrom<InterceptError<'a>> for Code {
-    type Error = InterceptError<'a>;
-
-    fn try_from(err: InterceptError<'a>) -> Result<Self, Self::Error> {
-        if let InterceptError::PulseError(code) = err {
-            Ok(code)
-        } else {
-            Err(err)
-        }
-    }
-}
-
-impl Display for InterceptError<'_> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), FormatError> {
-        match self {
-            Self::PulseError(code) => code.fmt(f),
-            Self::AtCapacity(limit, source) => write!(
-                f,
-                "Cannot intercept application {}, interceptor at limit ({})",
-                source.as_ref(),
-                limit
-            ),
-        }
-    }
-}
-
-/// Due to lifetime constraints, `InterceptError` cannot correctly implement
-/// [`Error::source()`]. To get a [`Code`] from an `InterceptError`, use
-/// [`Code::try_from`].
-impl Error for InterceptError<'_> {}
 
 // PUBLIC HELPER FUNCTIONS *****************************************************
 /// Primary entry point for all interceptors.
